New Daily Practice

+++ mysql/mySQL/backend/DoDeleteAction.php
@@ -1,12 +1,9 @@
 <?php
     $id = $_POST['id'];
-    $db_host = 'localhost';
-    $db_name = 'school';
-    $db_user = 'root';
-    $db_password = '1211';
-    $dsn="mysql:host=$db_host;dbname=$db_name;charset=utf8";
-    try {
-        $conn = new PDO($dsn,$db_user,$db_password);
+    require_once './mysql.inc.php';
+    $response = openDB();
+    if($response['status']==200){
+        $conn = $response['result'];
         $sql = "DELETE FROM `student` WHERE id=?";
         $stmt = $conn->prepare($sql);
         $result=$stmt->execute(array($id)); 
@@ -25,9 +22,6 @@
             $response['status']=400;
             $response['message']="SQL錯誤";
         }
-    } catch (PDOException $e) {
-        $response['status']=$e->getCode;
-        $response['message']=$e->getMessage();
     }
     echo json_encode($response);
 ?>
